requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
aiohttp==3.9.1
discord.py==2.3.2 